requirements.txt
numpy==1.19.2
pandas
matplotlib
Pillow
opencv-python-headless==4.5.1.48
plotly
requests
librosa==0.8.0
streamlit
tensorflow-cpu==2.4.1

